Use base vitest config
