+++ extensions/collections/vitest.config.ts
@@ -1,19 +1,24 @@
 import * as path from "node:path";
 
-import { defineConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
+
+import baseConfig from "../../vitest.base.config";
 
 const mock = (name: string) => path.resolve(import.meta.dirname, `__mocks__/${name}.ts`);
 
-export default defineConfig({
-  resolve: {
-    alias: {
-      "original-fs": "fs",
-      "@nexusmods/vortex-api": mock("vortex-api"),
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    resolve: {
+      alias: {
+        "original-fs": "fs",
+        "@nexusmods/vortex-api": mock("vortex-api"),
+      },
     },
-  },
-  test: {
-    environment: "happy-dom",
-    setupFiles: ["./test-setup.ts"],
-    include: ["src/**/*.test.ts"],
-  },
-});
+    test: {
+      environment: "happy-dom",
+      setupFiles: ["./test-setup.ts"],
+      include: ["src/**/*.test.ts"],
+    },
+  }),
+);
