+++ extensions/gamebryo-savegame-management/vitest.config.ts
@@ -1,15 +1,20 @@
 import * as path from "node:path";
 
-import { defineConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
 
-export default defineConfig({
-  resolve: {
-    alias: {
-      "@nexusmods/vortex-api": path.resolve(import.meta.dirname, "__mocks__/vortex-api.ts"),
+import baseConfig from "../../vitest.base.config";
+
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    resolve: {
+      alias: {
+        "@nexusmods/vortex-api": path.resolve(import.meta.dirname, "__mocks__/vortex-api.ts"),
+      },
     },
-  },
-  test: {
-    environment: "node",
-    include: ["src/**/*.test.ts", "test/**/*.test.ts"],
-  },
-});
+    test: {
+      environment: "node",
+      include: ["src/**/*.test.ts", "test/**/*.test.ts"],
+    },
+  }),
+);
