+++ extensions/games/game-baldursgate3/vitest.config.ts
@@ -1,11 +1,16 @@
-import { defineConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
 
-export default defineConfig({
-  test: {
-    environment: "node",
-    include: ["src/**/*.test.ts"],
-    // divine.exe invocations can be slow on cold caches.
-    testTimeout: 30000,
-    hookTimeout: 30000,
-  },
-});
+import baseConfig from "../../../vitest.base.config";
+
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    test: {
+      environment: "node",
+      include: ["src/**/*.test.ts"],
+      // divine.exe invocations can be slow on cold caches.
+      testTimeout: 30000,
+      hookTimeout: 30000,
+    },
+  }),
+);
