+++ extensions/games/game-xrebirth/vitest.config.ts
@@ -1,6 +1,8 @@
 import { createRequire } from "node:module";
 
-import { defineConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
+
+import baseConfig from "../../../vitest.base.config";
 
 // Resolve through Node's package resolution so the alias survives any future
 // move of the mock package. Regex `find` ensures `vortex-api` is the only
@@ -8,12 +10,15 @@ import { defineConfig } from "vitest/config";
 const require_ = createRequire(import.meta.url);
 const VORTEX_API_MOCK = require_.resolve("@vortex/extension-test-mocks");
 
-export default defineConfig({
-  resolve: {
-    alias: [{ find: /^@nexusmods\/vortex-api$/, replacement: VORTEX_API_MOCK }],
-  },
-  test: {
-    environment: "node",
-    include: ["src/**/*.test.ts"],
-  },
-});
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    resolve: {
+      alias: [{ find: /^@nexusmods\/vortex-api$/, replacement: VORTEX_API_MOCK }],
+    },
+    test: {
+      environment: "node",
+      include: ["src/**/*.test.ts"],
+    },
+  }),
+);
