+++ extensions/nmm-import-tool/vitest.config.ts
@@ -1,16 +1,21 @@
 import * as path from "node:path";
 
-import { defineConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
 
-export default defineConfig({
-  resolve: {
-    alias: {
-      "@nexusmods/vortex-api": path.resolve(import.meta.dirname, "__mocks__/vortex-api.ts"),
-      "modmeta-db": path.resolve(import.meta.dirname, "__mocks__/modmeta-db.ts"),
+import baseConfig from "../../vitest.base.config";
+
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    resolve: {
+      alias: {
+        "@nexusmods/vortex-api": path.resolve(import.meta.dirname, "__mocks__/vortex-api.ts"),
+        "modmeta-db": path.resolve(import.meta.dirname, "__mocks__/modmeta-db.ts"),
+      },
     },
-  },
-  test: {
-    environment: "happy-dom",
-    include: ["src/**/*.test.ts"],
-  },
-});
+    test: {
+      environment: "happy-dom",
+      include: ["src/**/*.test.ts"],
+    },
+  }),
+);
