+++ scripts/vitest.config.ts
@@ -1,9 +1,14 @@
-import { defineConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
 
-export default defineConfig({
-  test: {
-    name: "scripts",
-    environment: "node",
-    include: ["**/*.test.ts"],
-  },
-});
+import baseConfig from "../vitest.base.config";
+
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    test: {
+      name: "scripts",
+      environment: "node",
+      include: ["**/*.test.ts"],
+    },
+  }),
+);
