+++ src/main/vitest.config.ts
@@ -1,11 +1,14 @@
-import { defineConfig, type ViteUserConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
 
-const config: ViteUserConfig = defineConfig({
-  test: {
-    name: "@vortex/main",
-    environment: "node",
-    include: ["src/**/*.test.ts"],
-  },
-});
+import baseConfig from "../../vitest.base.config";
 
-export default config;
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    test: {
+      name: "@vortex/main",
+      environment: "node",
+      include: ["src/**/*.test.ts"],
+    },
+  }),
+);
