+++ src/main/vitest.downloader.config.ts
@@ -1,12 +1,15 @@
-import { defineConfig, type ViteUserConfig } from "vitest/config";
+import { mergeConfig, defineConfig } from "vitest/config";
 
-const config: ViteUserConfig = defineConfig({
-  test: {
-    name: "@vortex/main (downloader)",
-    environment: "node",
-    include: ["src/downloading/*.test.integration.ts"],
-    testTimeout: 30_000,
-  },
-});
+import baseConfig from "../../vitest.base.config";
 
-export default config;
+export default mergeConfig(
+  baseConfig,
+  defineConfig({
+    test: {
+      name: "@vortex/main (downloader)",
+      environment: "node",
+      include: ["src/downloading/*.test.integration.ts"],
+      testTimeout: 30_000,
+    },
+  }),
+);
