+++ tsconfig.node.json
@@ -8,6 +8,7 @@
     "./scripts/**/*.ts",
     "./ts-to-zod.config.mjs",
     "./vitest.config.ts",
+    "./vitest.base.config.ts",
     "./rolldown.base.mjs"
   ],
   "compilerOptions": {
