+++ vitest.config.ts
@@ -1,12 +1,8 @@
-import * as path from "node:path";
+import { mergeConfig } from "vitest/config";
 
-import { defineConfig } from "vitest/config";
+import baseConfig from "./vitest.base.config";
 
-const RESULTS_DIR = path.join(import.meta.dirname, "test-results");
-
-const isGitHubCI = process.env.CI && process.env.GITHUB_ACTIONS;
-
-export default defineConfig({
+export default mergeConfig(baseConfig, {
   test: {
     projects: [
       "./src/**/vitest.config.ts",
@@ -15,14 +11,5 @@ export default defineConfig({
       "./extensions/**/vitest.config.ts",
       "./scripts/vitest.config.ts",
     ],
-    reporters: ["default", "junit", isGitHubCI ? "github-actions" : undefined].filter(Boolean),
-    outputFile: {
-      junit: path.join(RESULTS_DIR, "junit.xml"),
-    },
-    coverage: {
-      provider: "v8",
-      reporter: ["text", "cobertura"],
-      reportsDirectory: path.join(RESULTS_DIR),
-    },
   },
 });
